requirements.txt
python-dotenv
mysql-connector-python
pyodbc
requests
flask
blinker
click
itsdangerous
jinja2
markupsafe
werkzeug
colorama
